rename some file and Class names to typical conventions
