+++ src/mindlessgen/symmetrization/__init__.py
@@ -2,14 +2,14 @@
 This module contains all structure modification classes.
 """
 
-from .StrucMod import StrucMod
+from .base import Symmetrizer
 from .mirror import Mirror
 from .inversion import Inversion
 from .rotation import CnRotation
 
 
 __all__ = [
-    "StrucMod",
+    "Symmetrizer",
     "Mirror",
     "Inversion",
     "CnRotation",
